New installation check improvement
Checks if the database "mysql" has been installed (i.e. the directory
"mysql" exists in the mysql data directory) instead of checking if
the directory is empty.

For example, the directory may not be empty if a dedicated filesystem
is used for the mysql data and thus a "lost+found" is present in the
directory:

    dd if=/dev/zero of=/tmp/mysql-data.img bs=1024 count=1048576
    mkfs.ext4 /tmp/mysql-data.img
    mount /tmp/mysql-data.img /mnt
    docker run -v /mnt:/var/lib/mysql -t mysql

Signed-off-by: Vincent Giersch <vincent.giersch@ovh.net>